Use dedicated Data Accessor class for attributes and features
